chmop($x) should suggest chomp
Text::Levenshtein does not take into account transpositions which results in
`chomp` not being suggested when `chmop($x)` is encountered.

That is:

    $ perl -Ilib -MDevel::DidYouMean -e 'chmop($x)'
    Undefined subroutine &main::chmop called at -e line 1.
    Did you mean chmod, chop?

This patch switches to `Text::Levenshtein::Damerau::XS` resulting in:

    $ perl -Ilib -MDevel::DidYouMean -e 'chmop($x)'
    Undefined subroutine &main::chmop called at -e line 1.
    Did you mean chop, chmod, or chomp?

The patch also incorporates a little beautification of the error message. As
always, beauty is in the eye of the beholder ;-)